Adds the possibility to implement Bots (Characters controlled by the AI)
This commit introduces the concept of **controlled by the server** on
the `NetworkedController`: When this property is set to `true` the
server collects the inputs and process the controlled object -> all the
other clients will see it as a doll.

Notice, it's possible to change `server_controlled` at runtime to switch
from player controller to AI controlled and vice versa.