Leave the optimization part for later.
For now, just call normal *_variant() and *_variant_size() functions.

Will be used for smaller packet size with often null variant(s) in the
data buffer.

NOTE: The properly non-delegated version of the optimized methods I
tried did not work well yet, it makes the character not sit on the
chair, and only set the player rotation to the chair seat, for some reasons.
So for now, it only calls the existing un-optimized version, so that we can use
the API from godotscript, and later separately properly implement it,
like when @andrea is back.