Fixes periodic hickups.
The hickups were caused by a bug into the function that calculated the
sub frames the client has to process on a given frame. Such function was
instructing the client to process 1500 frames on a single frame, causing
a massive hickup.
The bug was fixed, and the entire functionality that calculates the sub
frame count was moved into the SceneSync.

In addition, this commit, adds the possibilty to print std::string to
the standard out.