Fixes tests (#131)
When running the tests multiple time was possible to cause them to assert due some desync that happened because of the compensation algorithm was triggering random desync. These random desync make the tests miss behave that were unable to rely on some assumptions that allow them to verify the various algorithms.
The fix was essentially disabling the compensation algorithm (sub_ticks) feature entirely to ensure the test can trigger desync on purpose and verify all the algorithms like the recovery mechanism.

Notice, this bug isn't a regression. This was never triggered / detected before and in any case the bug was on the tests not on the code.