Refactors the FrameInput format so to include the buffer size.
The reason for this change was to make the NetSync much reliable. Since we can't guarantee that the client is ALWAYS uptodate we can't guarantee the input size calculation to work every time. With this change the size is included into the buffer therefore we made it more reliable and finally removed thecount_input_size which was very easy for the user to introduce bugs.